requirements.txt
flask
mysql
mysql-connector-python
flask-login
python-dotenv

